+++ cartographer/mapping_2d/pose_graph/constraint_builder.h
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-#ifndef CARTOGRAPHER_MAPPING_2D_SPARSE_POSE_GRAPH_CONSTRAINT_BUILDER_H_
-#define CARTOGRAPHER_MAPPING_2D_SPARSE_POSE_GRAPH_CONSTRAINT_BUILDER_H_
+#ifndef CARTOGRAPHER_MAPPING_2D_POSE_GRAPH_CONSTRAINT_BUILDER_H_
+#define CARTOGRAPHER_MAPPING_2D_POSE_GRAPH_CONSTRAINT_BUILDER_H_
 
 #include <array>
 #include <deque>
@@ -42,7 +42,7 @@
 
 namespace cartographer {
 namespace mapping_2d {
-namespace sparse_pose_graph {
+namespace pose_graph {
 
 // Returns (map <- submap) where 'submap' is a coordinate system at the origin
 // of the Submap.
@@ -178,8 +178,8 @@ class ConstraintBuilder {
   common::Histogram score_histogram_ GUARDED_BY(mutex_);
 };
 
-}  // namespace sparse_pose_graph
+}  // namespace pose_graph
 }  // namespace mapping_2d
 }  // namespace cartographer
 
-#endif  // CARTOGRAPHER_MAPPING_2D_SPARSE_POSE_GRAPH_CONSTRAINT_BUILDER_H_
+#endif  // CARTOGRAPHER_MAPPING_2D_POSE_GRAPH_CONSTRAINT_BUILDER_H_
